--- Category

+++ src/main/java/app/entity/Geo.java
@@ -16,16 +16,16 @@ public class Geo {
 
 
     @Column(name = "latitude")
-    private long latitude;
+    private double latitude;
 
     @Column(name = "longitude")
-    private long longitude;
+    private double longitude;
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
